fix: 로그인 로직 위치 롤백
